print ep0_data address
\0Usb device_init ep0_data=2216

+++ example/146_mch_mcu/source05/src/Uart.h
@@ -24,7 +24,7 @@ public:
     // vector
 
     fmt::format_to(std::back_inserter(ostr), fmt, std::forward<Args>(args)...);
-    SendString(ostr.data(), ostr.size());
+    SendString(ostr.data(), static_cast<uint16_t>(ostr.size()));
   }
 
   /** Overload for const char pointer
